chore: address PR comments
Issue-Id: #519
