+++ src/Geopilot.Api/Pipeline/IPipelineFactory.cs
@@ -17,7 +17,7 @@ public interface IPipelineFactory
     /// Creates a pipeline instance with the specified id.
     /// </summary>
     /// <param name="id">The id of the pipeline to be created.</param>
-    /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
+    /// <param name="cancellationToken">A cancellation token to cancel the pipeline during its run.</param>
     /// <returns>A <see cref="Pipeline"/> instance.</returns>
     /// <exception cref="Exception">Thrown when the pipeline cannot be created.</exception>
     IPipeline CreatePipeline(string id, CancellationToken cancellationToken);
