+++ src/Geopilot.Api/Program.cs
@@ -159,32 +159,7 @@ builder.Services.AddTransient<IFileProvider, PhysicalFileProvider>();
 builder.Services.AddTransient<IAssetHandler, AssetHandler>();
 builder.Services.AddHostedService<ValidationRunner>();
 builder.Services.AddHostedService<ValidationJobCleanupService>();
-
-Func<IServiceProvider, IPipelineFactory> cofigurePipelineFactory = (IServiceProvider sp) =>
-{
-    var configuration = sp.GetRequiredService<IConfiguration>();
-    var pipelineDefinition = configuration.GetValue<string>("Storage:Pipeline");
-    if (!string.IsNullOrWhiteSpace(pipelineDefinition))
-    {
-        var pipelineFactory = PipelineFactory.Builder()
-        .File(pipelineDefinition)
-        .Configuration(configuration)
-        .Build();
-
-        var validationErrors = pipelineFactory.PipelineProcessConfig.Validate();
-        if (validationErrors.HasErrors)
-        {
-            throw new InvalidOperationException($"errors in pipeline '{pipelineDefinition}': {validationErrors.ErrorMessage}");
-        }
-
-        return pipelineFactory;
-    }
-    else
-    {
-        throw new InvalidOperationException("unknown pipeline definition. define pipeline under 'Storage:Pipeline'.");
-    }
-};
-builder.Services.AddSingleton<IPipelineFactory>(cofigurePipelineFactory);
+builder.Services.RegisterPipelineFactory();
 
 builder.Services
     .AddHttpClient<IValidator, InterlisValidator>("INTERLIS_VALIDATOR_HTTP_CLIENT")
