+++ src/Geopilot.Api/appsettings.Development.json
@@ -30,8 +30,10 @@
   },
   "Storage": {
     "UploadDirectory": "Uploads",
-    "AssetsDirectory": "Persistent",
-    "Pipeline": "PipelineDefinitions\\basicPipeline_01.yaml"
+    "AssetsDirectory": "Persistent"
+  },
+  "Pipeline": {
+    "Definition": "PipelineDefinitions\\basicPipeline_01.yaml"
   },
   "Validation": {
     "InterlisCheckServiceUrl": "http://localhost:3080/"
