+++ src/Geopilot.Api/appsettings.json
@@ -23,7 +23,9 @@
   "Storage": {
     "UploadDirectory": "Uploads",
     "AssetsDirectory": "Persistent",
-    "Pipeline": ""
+  },
+  "Pipeline": {
+    "Definition": ""
   },
   "Validation": {
     "JobRetention": "1.00:00",
